WorkQueue.recv: handle task returned with no result file
There is a delay between when a task is marked as finished and complete.
If a task's execution is complete it is waiting for it's results to be returned to the master.
If it dies during this period, the task will return to the master's upper layer, and needs to be handled by application code.
In this case, we just resubmit it.